# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ERB System API Documentation",
        "description": "Tài liệu đặc tả và kiểm thử RESTful API - Hệ thống quản trị ERB",
        "contact": {
            "email": "admin@erb.vn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/transfer-orders": {
            "get": {
                "tags": [
                    "Transfer (Điều chuyển)"
                ],
                "summary": "Danh sách phiếu điều chuyển",
                "operationId": "9b72fcc22d2ded52560191c2b30d8464",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "warehouse_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transfer (Điều chuyển)"
                ],
                "summary": "Tạo phiếu điều chuyển (nháp)",
                "operationId": "7cfa689f5e7c58ac24119f151ca0a9d5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "source_warehouse_id",
                                    "dest_warehouse_id",
                                    "lines"
                                ],
                                "properties": {
                                    "source_warehouse_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "dest_warehouse_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "note": {
                                        "type": "string"
                                    },
                                    "lines": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "sku_id": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "quantity": {
                                                    "type": "number",
                                                    "example": 20
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo thành công"
                    },
                    "422": {
                        "description": "Kho nguồn trùng đích"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/transfer-orders/{id}": {
            "get": {
                "tags": [
                    "Transfer (Điều chuyển)"
                ],
                "summary": "Chi tiết phiếu điều chuyển",
                "operationId": "cbae0e922a8cd98f07c10068107e53c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/transfer-orders/{id}/submit": {
            "post": {
                "tags": [
                    "Transfer (Điều chuyển)"
                ],
                "summary": "Gửi duyệt",
                "operationId": "07f3b63d59bf3a63f8651f23def4ba67",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/transfer-orders/{id}/approve": {
            "post": {
                "tags": [
                    "Transfer (Điều chuyển)"
                ],
                "summary": "Phê duyệt (kiểm tra tồn nguồn)",
                "operationId": "19248f4851fabf02bfc8b01d4c6c9263",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Tồn nguồn không đủ"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/transfer-orders/{id}/dispatch": {
            "post": {
                "tags": [
                    "Transfer (Điều chuyển)"
                ],
                "summary": "Xuất điều chuyển (nguồn -, In-Transit +)",
                "operationId": "d8e268d2d0e06d88712cbf0db9a2ded7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Đang vận chuyển"
                    },
                    "422": {
                        "description": "Tồn không đủ"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/transfer-orders/{id}/receive": {
            "post": {
                "tags": [
                    "Transfer (Điều chuyển)"
                ],
                "summary": "Nhận hàng (In-Transit -, đích +)",
                "operationId": "5feed0b890ff2d1156e36cdc02b8506c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "lines": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "line_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "received_qty": {
                                                    "type": "number",
                                                    "example": 20
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Hoàn tất"
                    },
                    "422": {
                        "description": "Nhận vượt số điều chuyển"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/transfer-orders/{id}/return": {
            "post": {
                "tags": [
                    "Transfer (Điều chuyển)"
                ],
                "summary": "Trả lại điều chuyển (hoàn nguồn)",
                "operationId": "7ab6ec6dbe676fd4658770272437e14f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bị trả lại"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/transfer-orders/{id}/cancel": {
            "post": {
                "tags": [
                    "Transfer (Điều chuyển)"
                ],
                "summary": "Hủy phiếu (trước khi xuất)",
                "operationId": "9afdd2f7ae9e8aeeb1ca890d226d0f26",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Nhập JWT Access Token (dạng: Bearer <token>) để thực hiện các yêu cầu cần xác thực.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}